tests/wpt/web-platform-tests/css/motion/parsing/offset-path-parsing-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Motion Path Module Level 1: parsing offset-path with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/motion-1/#offset-path-property">
<meta name="assert" content="offset-path supports the full grammar from the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("offset-path", "none");

test_valid_value("offset-path", "ray(0rad closest-side)");
test_valid_value("offset-path", "ray(0.25turn closest-corner contain)");
test_valid_value("offset-path", "ray(200grad farthest-side)");
test_valid_value("offset-path", "ray(270deg farthest-corner contain)");
test_valid_value("offset-path", "ray(-720deg sides)");
test_valid_value("offset-path", "ray(calc(180deg - 45deg) farthest-side)", "ray(calc(135deg) farthest-side)");

test_valid_value("offset-path", "path('m 0 0 h -100')");
test_valid_value("offset-path", "path('M 0 0 L 100 100 M 100 200 L 200 200 Z L 300 300 Z')");

test_valid_value("offset-path", 'url("http://www.example.com/index.html#polyline1")');

test_valid_value("offset-path", "circle(100px)");
test_valid_value("offset-path", "margin-box");
test_valid_value("offset-path", "inset(10% 20% 30% 40%) border-box");
test_valid_value("offset-path", "fill-box ellipse(50% 60%)", "ellipse(50% 60%) fill-box");
</script>
</body>
</html>
